weather_map/requirements.txt
Pillow==2.2.1
selenium==3.141.0
